requirements.txt
boto3
numpy
opencv-python-headless  
Pillow
torch
facenet-pytorch
imutils

